+++ Docker/SIGAC/app/Http/Controllers/CategoriaController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Repositories\CategoriaRepository;
 use App\Repositories\CursoRepository;
 use App\Models\Categoria;
@@ -16,17 +17,22 @@ class CategoriaController extends Controller {
     }
 
     public function index() {
-        $data = $this->repository->selectAllWith(['curso']);
+
+        $this->authorize('hasFullPermission', Categoria::class);
+        $data = $this->repository->findByColumnWith('curso_id', Auth::user()->curso_id, ['curso']);
         return view('categoria.index', compact('data'));    
     }
 
     public function create() {
+
+        $this->authorize('hasFullPermission', Categoria::class);
         $cursos = (new CursoRepository())->selectAll();
         return view('categoria.create', compact(['cursos']));
     }
 
     public function store(Request $request) {
         
+        $this->authorize('hasFullPermission', Categoria::class);
         $objCurso = (new CursoRepository())->findById($request->curso_id);
         
         if(isset($objCurso)) {
@@ -48,6 +54,7 @@ class CategoriaController extends Controller {
 
     public function show(string $id) {
 
+        $this->authorize('hasFullPermission', Categoria::class);
         $cursos = (new CursoRepository())->selectAll();
         $data = $this->repository->findByIdWith(['curso'], $id);
         if(isset($data))
@@ -63,6 +70,7 @@ class CategoriaController extends Controller {
 
     public function edit(string $id) {
         
+        $this->authorize('hasFullPermission', Categoria::class);
         $data = $this->repository->findById($id);
         if(isset($data)) {
             $cursos = (new CursoRepository())->selectAll();
@@ -79,6 +87,7 @@ class CategoriaController extends Controller {
 
     public function update(Request $request, string $id) {
 
+        $this->authorize('hasFullPermission', Categoria::class);
         $obj = $this->repository->findById($id);
         $objCurso = (new CursoRepository())->findById($request->curso_id);
         
@@ -100,6 +109,7 @@ class CategoriaController extends Controller {
 
     public function destroy(string $id) {
         
+        $this->authorize('hasFullPermission', Categoria::class);
         if($this->repository->delete($id))  {
             return redirect()->route('categoria.index');
         }
